Give credit to photo + change the button link
I suggest you to make this a landing page instead of a home page because sometimes the user will want to click home to go back to a "main page".
When the user is logged in, and the user click the "Start Shopping" button, the user will always redirect to the login page
Maybe do a conditional statement for the redirect, or on the login page.

Also, always give credit to photo taken from the internet. I was able to search the image on Google.